enroll: remove `NewParticipant` class
I initially though this would allow us to differentiate between the
old participant and enrolled participant. But we need to impl the sign
functionalities of old participant in `NewParticipant` again.

So, better to use the same class for both old and enrolled participant
and simply differntiate them in the comments